+++ src/greek_bess/__init__.py
@@ -1,3 +1,3 @@
 """Greek Battery Investment Stress Tester."""
 
-__version__ = "0.8.2"
+__version__ = "0.8.3"